WIP: Shuffle everything around
