+++ packages/desktop/app/components/PatientPrinting/index.js
@@ -1,3 +1,12 @@
-export * from './PrintPortal';
-export * from './PrintLetterhead';
-export * from './PatientPrintDetailsModal';
+export * from './modals/PatientPrintDetailsModal';
+export * from './modals/PrintMultipleMedicationSelectionModal';
+export * from './modals/PrintMultipleLabRequestsSelectionModal';
+export * from './modals/PrintMultipleImagingRequestsSelectionModal';
+export * from './modals/PrescriptionPrintModal';
+export * from './modals/ImmunisationCertificateModal';
+
+export * from './printouts/LabRequestPrintout';
+export * from './printouts/ImagingRequestPrintout';
+
+export { PrintLetterhead } from './printouts/reusable/PrintLetterhead';
+export { PrintPortal } from './printouts/reusable/PrintPortal';
