+++ packages/desktop/app/components/PatientPrinting/modals/MultipleLabRequestsPrintoutModal.js
@@ -2,36 +2,26 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { useQuery } from '@tanstack/react-query';
 
-import { Modal } from '../Modal';
-import { MultiplePrescriptionPrintout } from './MultiplePrescriptionPrintout';
-import { LoadingIndicator } from '../LoadingIndicator';
-import { useCertificate } from '../../utils/useCertificate';
-import { useApi } from '../../api';
-import { Colors } from '../../constants';
+import { Modal } from '../../Modal';
+import { LoadingIndicator } from '../../LoadingIndicator';
+import { useCertificate } from '../../../utils/useCertificate';
+import { useApi } from '../../../api';
+import { Colors } from '../../../constants';
 
-export const MultiplePrescriptionPrintoutModal = ({
-  encounter,
-  prescriberId,
-  prescriptions,
-  open,
-  onClose,
-}) => {
+import { MultipleLabRequestsPrintout } from '../printouts/MultipleLabRequestsPrintout';
+
+export const MultipleLabRequestsPrintoutModal = ({ encounter, labRequests, open, onClose }) => {
   const certificateData = useCertificate();
   const api = useApi();
 
   const { data: patient, isLoading: patientLoading } = useQuery(
     ['patient', encounter.patientId],
-    () => api.get(`patient/${encounter.patientId}`),
-  );
-
-  const { data: prescriber, isLoading: prescriberLoading } = useQuery(
-    ['prescriber', prescriberId],
-    () => api.get(`user/${prescriberId}`),
+    () => api.get(`patient/${encodeURIComponent(encounter.patientId)}`),
   );
 
   const { data: additionalData, isLoading: additionalDataLoading } = useQuery(
     ['additionalData', encounter.patientId],
-    () => api.get(`patient/${encounter.patientId}/additionalData`),
+    () => api.get(`patient/${encodeURIComponent(encounter.patientId)}/additionalData`),
   );
 
   const { data: village = {}, isLoading: villageQueryLoading } = useQuery(
@@ -46,31 +36,29 @@ export const MultiplePrescriptionPrintoutModal = ({
 
   return (
     <Modal
-      title="Print prescriptions"
+      title="Print lab requests"
       width="md"
       open={open}
       onClose={onClose}
       color={Colors.white}
       printable
     >
-      {patientLoading || additionalDataLoading || villageLoading || prescriberLoading ? (
+      {patientLoading || additionalDataLoading || villageLoading ? (
         <LoadingIndicator />
       ) : (
-        <MultiplePrescriptionPrintout
+        <MultipleLabRequestsPrintout
           certificateData={certificateData}
           patientData={{ ...patient, additionalData, village }}
-          prescriber={prescriber}
-          prescriptions={prescriptions}
+          labRequests={labRequests}
         />
       )}
     </Modal>
   );
 };
 
-MultiplePrescriptionPrintoutModal.propTypes = {
+MultipleLabRequestsPrintoutModal.propTypes = {
   encounter: PropTypes.object.isRequired,
-  prescriberId: PropTypes.string.isRequired,
-  prescriptions: PropTypes.array.isRequired,
+  labRequests: PropTypes.array.isRequired,
   open: PropTypes.bool.isRequired,
   onClose: PropTypes.func.isRequired,
 };
