+++ packages/desktop/app/components/PatientPrinting/modals/PrintMultipleImagingRequestsSelectionForm.js
@@ -2,16 +2,17 @@ import React, { useState, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { useQuery } from '@tanstack/react-query';
 
-import { Table, useSelectableColumn } from '../Table';
-import { OuterLabelFieldWrapper } from '../Field';
-import { ConfirmCancelRow } from '../ButtonRow';
-import { DateDisplay } from '../DateDisplay';
+import { Table, useSelectableColumn } from '../../Table';
+import { OuterLabelFieldWrapper } from '../../Field';
+import { ConfirmCancelRow } from '../../ButtonRow';
+import { DateDisplay } from '../../DateDisplay';
+import { useApi } from '../../../api';
+import { Colors } from '../../../constants';
+import { getImagingRequestType } from '../../../utils/getImagingRequestType';
+import { getAreaNote } from '../../../utils/areaNote';
+import { useLocalisation } from '../../../contexts/Localisation';
+
 import { MultipleImagingRequestsPrintoutModal } from './MultipleImagingRequestsPrintoutModal';
-import { useApi } from '../../api';
-import { Colors } from '../../constants';
-import { getImagingRequestType } from '../../utils/getImagingRequestType';
-import { getAreaNote } from '../../utils/areaNote';
-import { useLocalisation } from '../../contexts/Localisation';
 
 const COLUMN_KEYS = {
   ID: 'id',
