+++ packages/desktop/app/components/PatientPrinting/modals/PrintMultipleLabRequestsSelectionModal.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Modal } from '../Modal';
+import { Modal } from '../../Modal';
 import { PrintMultipleLabRequestsSelectionForm } from './PrintMultipleLabRequestsSelectionForm';
 
 export const PrintMultipleLabRequestsSelectionModal = ({ encounter, open, onClose }) => {
