+++ packages/desktop/app/components/PatientPrinting/printouts/DeathCertificate.js
@@ -2,13 +2,13 @@ import React from 'react';
 import styled from 'styled-components';
 import { Typography, Box } from '@material-ui/core';
 import { getCurrentDateTimeString } from 'shared/utils/dateTime';
-import { PrintLetterhead } from './PrintLetterhead';
-import { DateDisplay } from '../DateDisplay';
+import { PrintLetterhead } from './reusable/PrintLetterhead';
+import { DateDisplay } from '../../DateDisplay';
 import {
   LocalisedCertificateLabel as LocalisedLabel,
   CertificateLabel as Label,
-} from './CertificateLabels';
-import { CertificateWrapper } from './CertificateWrapper';
+} from './reusable/CertificateLabels';
+import { CertificateWrapper } from './reusable/CertificateWrapper';
 
 const Grid = styled(Box)`
   display: grid;
