+++ packages/desktop/app/components/PatientPrinting/printouts/MultipleLabRequestsPrintout.js
@@ -5,16 +5,17 @@ import Divider from '@material-ui/core/Divider';
 
 import { getCurrentDateString } from 'shared/utils/dateTime';
 
-import { LocalisedLabel } from './SimplePrintout';
-import { DateDisplay } from '../DateDisplay';
-import { PrintLetterhead } from './PrintLetterhead';
-import { CertificateWrapper } from './CertificateWrapper';
-import { ListTable } from './ListTable';
-import { PatientDetailPrintout } from './PatientDetailPrintout';
-import { CertificateLabel } from './CertificateLabels';
-import { useAuth } from '../../contexts/Auth';
-import { Colors } from '../../constants';
-import { NotesPagesSection } from './NotesPagesSection';
+import { DateDisplay } from '../../DateDisplay';
+import { useAuth } from '../../../contexts/Auth';
+import { Colors } from '../../../constants';
+
+import { LocalisedLabel } from './reusable/SimplePrintout';
+import { PrintLetterhead } from './reusable/PrintLetterhead';
+import { CertificateWrapper } from './reusable/CertificateWrapper';
+import { ListTable } from './reusable/ListTable';
+import { PatientDetailPrintout } from './reusable/PatientDetailPrintout';
+import { CertificateLabel } from './reusable/CertificateLabels';
+import { NotesPagesSection } from './reusable/NotesPagesSection';
 
 const RowContainer = styled.div`
   display: flex;
