+++ packages/desktop/app/components/PatientPrinting/printouts/reusable/PatientBarcode.js
