+++ packages/desktop/app/components/PatientPrinting/printouts/reusable/PatientIDCardPage.js
@@ -1,9 +1,11 @@
 import React, { useEffect } from 'react';
 import styled from 'styled-components';
-import { useLocalisation } from '../../contexts/Localisation';
-import { useElectron } from '../../contexts/Electron';
-import { SEX_VALUE_INDEX } from '../../constants';
-import { DateDisplay } from '../DateDisplay';
+
+import { useLocalisation } from '../../../../contexts/Localisation';
+import { useElectron } from '../../../../contexts/Electron';
+import { SEX_VALUE_INDEX } from '../../../../constants';
+import { DateDisplay } from '../../../DateDisplay';
+
 import { PatientBarcode } from './PatientBarcode';
 import { PrintPortal } from './PrintPortal';
