+++ packages/desktop/stories/certificate.stories.js
@@ -4,8 +4,8 @@ import { storiesOf } from '@storybook/react';
 import { createDummyPatient, createDummyPatientAdditionalData } from 'shared/demoData';
 import { CovidLabCertificate, VaccineCertificate } from 'shared/utils/patientCertificates';
 import { PDFViewer } from '@react-pdf/renderer';
-import { DeathCertificate } from '../app/components/PatientPrinting/DeathCertificate';
-import { LabRequestPrintout } from '../app/components/PatientPrinting/LabRequestPrintout';
+import { DeathCertificate } from '../app/components/PatientPrinting/printouts/DeathCertificate';
+import { LabRequestPrintout } from '../app/components/PatientPrinting';
 import SigningImage from './assets/signing-image.png';
 import Watermark from './assets/watermark.png';
 import Logo from './assets/tamanu-logo.png';
